RSSEC-29 feat: add validation function in method valid() of the class InnerForm, add some needed actions in method create()
